creating explore header component
